Chromium updater script

+++ reinstall_c5.sh
@@ -4,17 +4,21 @@
 REPO_PATH=/Library/WebServer/Documents/concrete5
 
 echo -e "\n[+] Reinstalling local concrete5 instance..."
-cd $REPO_PATH
 
-echo -e "[+] updating files from GitHub repository..."
-git pull origin develop
+if [ -d $REPO_PATH ]
+then
+    cd $REPO_PATH
 
-echo -e "[+] removing temporary files..."
-rm -rf application/cache/*
-rm -rf application/files/*
+    echo -e "[+] updating files from GitHub repository..."
+    git pull origin develop
 
-echo -e "[+] deleting database.php..."
-cd $REPO_PATH/concrete/config
-rm -f database.php
+    echo -e "[+] removing temporary files..."
+    rm -rf $REPO_PATH/application/cache/*
+    rm -rf $REPO_PATH/application/files/*
 
-echo -e "\n[+] DONE"
+    echo -e "[+] deleting database.php..."
+    cd $REPO_PATH/concrete/config
+    rm -f database.php
+
+    echo -e "\n[+] DONE"
+fi
